Replace Semigroup with Semi
