+++ src/semigroup.rs
@@ -15,7 +15,7 @@
 //!
 //! let expected = (2, 5.0f32, String::from("hi world"), Some(3));
 //!
-//! assert_eq!(t1.combine(&t2), expected);
+//! assert_eq!(t1.combine(t2), expected);
 //!
 //! // ultimately, the Tuple-based Semigroup implementations are only available for a maximum of
 //! // 26 elements. If you need more, use HList, which is has no such limit.
@@ -23,77 +23,109 @@
 //! let h1 = hlist![1, 3.3, 53i64];
 //! let h2 = hlist![2, 1.2, 1i64];
 //! let h3 = hlist![3, 4.5, 54];
-//! assert_eq!(h1.combine(&h2), h3)
+//! assert_eq!(h1.combine(h2), h3)
 //! # }
 //! ```
 
+use frunk_core::hlist::*;
+use std::ops::{Deref, BitAnd, BitOr};
+use std::borrow::Borrow;
+use std::cmp::Ordering;
 use std::cell::*;
 use std::hash::Hash;
-use std::ops::{Deref, BitAnd, BitOr};
-use std::cmp::Ordering;
 use std::collections::{HashSet, HashMap};
 use std::collections::hash_map::Entry;
-use frunk_core::hlist::*;
 
 /// Wrapper type for types that are ordered and can have a Max combination
 #[derive(PartialEq, Debug, Eq, Clone, Copy, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "with_serde", derive(Serialize, Deserialize))]
 pub struct Max<T: Ord>(pub T);
 
 /// Wrapper type for types that are ordered and can have a Min combination
 #[derive(PartialEq, Debug, Eq, Clone, Copy, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "with_serde", derive(Serialize, Deserialize))]
 pub struct Min<T: Ord>(pub T);
 
 /// Wrapper type for types that can have a Product combination
 #[derive(PartialEq, Debug, Eq, Clone, Copy, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "with_serde", derive(Serialize, Deserialize))]
 pub struct Product<T>(pub T);
 
 /// Wrapper type for boolean that acts as a bitwise && combination
 #[derive(PartialEq, Debug, Eq, Clone, Copy, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "with_serde", derive(Serialize, Deserialize))]
 pub struct All<T>(pub T);
 
 /// Wrapper type for boolean that acts as a bitwise || combination
 #[derive(PartialEq, Debug, Eq, Clone, Copy, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "with_serde", derive(Serialize, Deserialize))]
 pub struct Any<T>(pub T);
 
-/// A Semigroup is a class of thing that has a definable combine operation
-pub trait Semigroup {
-    /// Associative operation taking which combines two values.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use frunk::semigroup::*;
-    /// assert_eq!(Some(1).combine(&Some(2)), Some(3))
-    /// ```
-    fn combine(&self, other: &Self) -> Self;
+/// Output is a parameter because we want to allow Semigroup to not
+/// necessarily return Self (e.g. in the case of Self being a pointer)
+///
+/// RHS is also a parameter in case we need to chain more combinations together..
+///
+/// Also, using a type parameter instead of an associated type because
+/// there was a weird diverging trait search going with an associated type.
+///
+/// This means that yes, we need to enforce things with Laws.
+pub trait Semigroup<Output = Self, RHS = Self> {
+    fn combine(self, other: RHS) -> Output;
 }
+
 
 /// Allow the combination of any two HLists having the same structure
 /// if all of the sub-element types are also Semiups
-impl<H: Semigroup, T: HList + Semigroup> Semigroup for HCons<H, T> {
-    fn combine(&self, other: &Self) -> Self {
-        self.tail.combine(&other.tail).prepend(
-            self.head.combine(&other.head),
-        )
+impl<H, HO, T, TO> Semigroup<HCons<HO, TO>> for HCons<H, T>
+where
+    H: Semigroup<HO>,
+    T: HList + Semigroup<TO>,
+{
+    fn combine(self, other: Self) -> HCons<HO, TO> {
+        let tail_comb = self.tail.combine(other.tail);
+        let h_comb = self.head.combine(other.head);
+        HCons {
+            head: h_comb,
+            tail: tail_comb,
+        }
     }
 }
 
 /// Since () + () = (), the same is true for HNil
 impl Semigroup for HNil {
-    fn combine(&self, _: &Self) -> Self {
-        *self
+    fn combine(self, _: Self) -> Self {
+        self
+    }
+}
+
+impl<T, TO> Semigroup<Option<TO>> for Option<T>
+where
+    T: Semigroup<TO>,
+    TO: From<T>,
+{
+    fn combine(self, other: Self) -> Option<TO> {
+        if let Some(s) = self {
+            if let Some(o) = other {
+                Some(s.combine(o))
+            } else {
+                Some(TO::from(s))
+            }
+        } else {
+            other.map(TO::from)
+        }
     }
 }
 
 /// Return this combined with itself `n` times.
-pub fn combine_n<T>(o: &T, times: u32) -> T
+pub fn combine_n<T>(o: T, times: u32) -> T
 where
     T: Semigroup + Clone,
 {
-    let mut x = o.clone();
     // note: range is non-inclusive in the upper bound
+    let mut x = o.clone();
     for _ in 1..times {
-        x = o.combine(&x);
+        x = x.combine(o.clone());
     }
     x
 }
@@ -106,178 +138,112 @@ where
 ///
 /// ```
 /// # use frunk::semigroup::*;
-/// let v1 = &vec![1, 2, 3];
+/// let v1 = vec![1, 2, 3];
 /// assert_eq!(combine_all_option(v1), Some(6));
 ///
 /// let v2: Vec<i16> = Vec::new(); // empty!
-/// assert_eq!(combine_all_option(&v2), None);
+/// assert_eq!(combine_all_option(v2), None);
 /// ```
-pub fn combine_all_option<T>(xs: &Vec<T>) -> Option<T>
+pub fn combine_all_option<T>(mut xs: Vec<T>) -> Option<T>
 where
-    T: Semigroup + Clone,
+    T: Semigroup,
 {
-    match xs.first() {
-        Some(ref head) => {
-            let tail = xs[1..].to_vec();
-            // TODO figure out how to write this as a fold
-            let mut x = (*head).clone();
-            for i in tail {
-                x = x.combine(&i)
+    if xs.len() < 1 {
+        None
+    } else {
+        // Chop off xs
+        let tail = xs.split_off(1);
+        xs.pop().map(|mut x| {
+            for i in tail.into_iter() {
+                x = x.combine(i)
             }
-            Some(x)
-        }
-        _ => None,
+            x
+        })
     }
 }
 
-macro_rules! numeric_semigroup_imps {
+macro_rules! numeric_semi_imps {
   ($($tr:ty),*) => {
     $(
       impl Semigroup for $tr {
-        fn combine(&self, other: &Self) -> Self { self + other }
+        fn combine(self, other: Self) -> $tr { self + other }
+      }
+      impl <'a> Semigroup<$tr, &'a $tr> for $tr {
+        fn combine(self, other: &'a $tr) -> $tr { self + other }
+      }
+      impl <'a> Semigroup<$tr> for &'a $tr {
+        fn combine(self, other: Self) -> $tr { self + other }
       }
     )*
   }
 }
 
-numeric_semigroup_imps!(i8, i16, i32, i64, u8, u16, u32, u64, isize, usize, f32, f64);
+numeric_semi_imps!(i8, i16, i32, i64, u8, u16, u32, u64, isize, usize, f32, f64);
 
-macro_rules! numeric_product_semigroup_imps {
-  ($($tr:ty),*) => {
-    $(
-      impl Semigroup for Product<$tr> {
-        fn combine(&self, other: &Self) -> Self {
-            let Product(x) = *self;
-            let Product(y) = *other;
-            Product(x * y)
-         }
-      }
-    )*
-  }
+impl<Str: Borrow<str>> Semigroup<String, Str> for String {
+    fn combine(self, other: Str) -> Self {
+        let mut s = self;
+        s.push_str(other.borrow());
+        s
+    }
 }
 
-numeric_product_semigroup_imps!(i8, i16, i32, i64, u8, u16, u32, u64, isize, usize, f32, f64);
+impl<'a, Str: Borrow<str>> Semigroup<String, Str> for &'a str {
+    fn combine(self, other: Str) -> String {
+        let s = self.to_string();
+        s.combine(other)
+    }
+}
 
-impl<T> Semigroup for Option<T>
+impl<T, TO> Semigroup<Box<TO>> for Box<T>
 where
-    T: Semigroup + Clone,
+    T: Semigroup<TO>,
 {
-    fn combine(&self, other: &Self) -> Self {
-        match *self {
-            Some(ref v) => {
-                match *other {
-                    Some(ref v_other) => Some(v.combine(v_other)),
-                    _ => self.clone(),
-                }
-            }
-            _ => other.clone(),
-        }
+    fn combine(self, other: Self) -> Box<TO> {
+        let s = *self;
+        let o = *other;
+        Box::new(s.combine(o))
     }
 }
 
-impl<T: Semigroup> Semigroup for Box<T> {
-    fn combine(&self, other: &Self) -> Self {
-        Box::new(self.deref().combine(other.deref()))
+impl<'a, T, TO> Semigroup<Box<TO>> for &'a Box<T>
+where
+    &'a T: Semigroup<TO>,
+{
+    fn combine(self, other: Self) -> Box<TO> {
+        let s = self.deref();
+        let o = other.deref();
+        Box::new(s.combine(o))
     }
 }
 
-impl Semigroup for String {
-    fn combine(&self, other: &Self) -> Self {
-        let mut cloned = self.clone();
-        cloned.push_str(other);
-        cloned
-    }
-}
-
-impl<T: Clone> Semigroup for Vec<T> {
-    fn combine(&self, other: &Self) -> Self {
-        let mut v = self.clone();
-        v.extend_from_slice(other);
+impl<T> Semigroup for Vec<T> {
+    fn combine(self, other: Self) -> Self {
+        let mut v = self;
+        let mut o = other;
+        v.append(&mut o);
         v
-    }
-}
-
-impl<T> Semigroup for Cell<T>
-where
-    T: Semigroup + Copy,
-{
-    fn combine(&self, other: &Self) -> Self {
-        Cell::new(self.get().combine(&(other.get())))
-    }
-}
-
-impl<T: Semigroup> Semigroup for RefCell<T> {
-    fn combine(&self, other: &Self) -> Self {
-        let self_b = self.borrow();
-        let other_b = other.borrow();
-        RefCell::new(self_b.deref().combine(other_b.deref()))
-    }
-}
-
-impl<T> Semigroup for HashSet<T>
-where
-    T: Eq + Hash + Clone,
-{
-    fn combine(&self, other: &Self) -> Self {
-        let mut h = HashSet::new();
-        for i in self {
-            h.insert(i.clone());
-        }
-        for i in other {
-            h.insert(i.clone());
-        }
-        h
-    }
-}
-
-impl<K, V> Semigroup for HashMap<K, V>
-where
-    K: Eq + Hash + Clone,
-    V: Semigroup + Clone,
-{
-    fn combine(&self, other: &Self) -> Self {
-        let mut h: HashMap<K, V> = HashMap::new();
-        for (k, v) in self {
-            h.insert(k.clone(), v.clone());
-        }
-        for (k, v) in other {
-            let k_clone = k.clone();
-            match h.entry(k_clone) {
-                Entry::Occupied(o) => {
-                    let existing = o.into_mut();
-                    let comb = existing.combine(&v);
-                    *existing = comb;
-                }
-                Entry::Vacant(o) => {
-                    o.insert(v.clone());
-                }
-            }
-        }
-        h
     }
 }
 
 impl<T> Semigroup for Max<T>
 where
-    T: Ord + Clone,
+    T: Ord,
 {
-    fn combine(&self, other: &Self) -> Self {
-        let x = self.0.clone();
-        let y = other.0.clone();
+    fn combine(self, Max(y): Self) -> Self {
+        let Max(x) = self;
         match x.cmp(&y) {
             Ordering::Less => Max(y),
             _ => Max(x),
         }
     }
 }
-
 impl<T> Semigroup for Min<T>
 where
-    T: Ord + Clone,
+    T: Ord,
 {
-    fn combine(&self, other: &Self) -> Self {
-        let x = self.0.clone();
-        let y = other.0.clone();
+    fn combine(self, Min(y): Self) -> Self {
+        let Min(x) = self;
         match x.cmp(&y) {
             Ordering::Less => Min(x),
             _ => Min(y),
@@ -290,7 +256,7 @@ macro_rules! simple_all_impls {
     ($($tr:ty)*) => {
         $(
             impl Semigroup for All<$tr> {
-                fn combine(&self, other: &Self) -> Self {
+                fn combine(self, other: Self) -> Self {
                     let x = self.0;
                     let y = other.0;
                     All(x.bitand(y))
@@ -302,11 +268,12 @@ macro_rules! simple_all_impls {
 
 simple_all_impls! { bool usize u8 u16 u32 u64 isize i8 i16 i32 i64 }
 
+
 macro_rules! simple_any_impls {
     ($($tr:ty)*) => {
         $(
             impl Semigroup for Any<$tr> {
-                fn combine(&self, other: &Self) -> Self {
+                fn combine(self, other: Self) -> Self {
                     let x = self.0;
                     let y = other.0;
                     Any(x.bitor(y))
@@ -317,6 +284,92 @@ macro_rules! simple_any_impls {
 }
 
 simple_any_impls! { bool usize u8 u16 u32 u64 isize i8 i16 i32 i64 }
+
+
+macro_rules! numeric_product_semigroup_imps {
+  ($($tr:ty),*) => {
+    $(
+      impl Semigroup for Product<$tr> {
+        fn combine(self, other: Self) -> Self {
+            let Product(x) = self;
+            let Product(y) = other;
+            Product(x * y)
+         }
+      }
+    )*
+  }
+}
+
+numeric_product_semigroup_imps!(i8, i16, i32, i64, u8, u16, u32, u64, isize, usize, f32, f64);
+
+impl<T> Semigroup for Cell<T>
+where
+    T: Semigroup + Copy,
+{
+    fn combine(self, other: Self) -> Self {
+        Cell::new(self.get().combine((other.get())))
+    }
+}
+
+impl<T, Out> Semigroup<RefCell<Out>> for RefCell<T>
+where
+    T: ToOwned,
+    <T as ToOwned>::Owned: Semigroup<Out>,
+{
+    fn combine(self, other: Self) -> RefCell<Out> {
+        let self_b = self.borrow().deref().to_owned();
+        let other_b = other.borrow().to_owned();
+        RefCell::new(self_b.combine(other_b))
+    }
+}
+
+impl<T> Semigroup for HashSet<T>
+where
+    T: Eq + Hash,
+{
+    fn combine(self, other: Self) -> Self {
+        let mut h = HashSet::new();
+        for i in self {
+            h.insert(i);
+        }
+        for i in other {
+            h.insert(i);
+        }
+        h
+    }
+}
+
+impl<K, V> Semigroup for HashMap<K, V>
+where
+    K: Eq + Hash,
+    V: Semigroup,
+{
+    fn combine(self, other: Self) -> Self {
+        let mut h: HashMap<K, V> = HashMap::new();
+        for (k, v) in self {
+            h.insert(k, v);
+        }
+        let mut combined = vec![];
+        for (k, v) in other {
+            match h.entry(k) {
+                Entry::Occupied(o) => {
+                    // Store and insert later
+                    let (k, existing) = o.remove_entry();
+                    let comb = existing.combine(v);
+                    combined.push((k, comb));
+                }
+                Entry::Vacant(o) => {
+                    o.insert(v);
+                }
+            }
+        }
+        for (k, v) in combined {
+            h.insert(k, v);
+        }
+        h
+    }
+}
+
 
 macro_rules! tuple_impls {
     () => {}; // no more
@@ -339,8 +392,8 @@ macro_rules! tuple_impls {
 // Finally expand into our implementation
     ([($idx:tt, $typ:ident); $( ($nidx:tt, $ntyp:ident); )*]) => {
         impl<$typ: Semigroup, $( $ntyp: Semigroup),*> Semigroup for ($typ, $( $ntyp ),*) {
-            fn combine(&self, other: &Self) -> Self {
-                (self.$idx.combine(&other.$idx), $(self.$nidx.combine(&other.$nidx), )*)
+            fn combine(self, other: Self) -> Self {
+                (self.$idx.combine(other.$idx), $(self.$nidx.combine(other.$nidx), )*)
             }
         }
     }
@@ -374,7 +427,7 @@ tuple_impls! {
 mod tests {
     use super::*;
 
-    macro_rules! semigroup_tests {
+    macro_rules! semi_tests {
       ($($name:ident, $comb: expr => $expected: expr, $tr:ty)+) => {
         $(
           #[test]
@@ -386,42 +439,116 @@ mod tests {
       }
     }
 
-    semigroup_tests! {
-        test_i8, 1.combine(&2) => 3, i8
-        test_product_i8, Product(1).combine(&Product(2)) => Product(2), Product<i8>
-        test_i16, 1.combine(&2) => 3, i16
-        test_i32, 1.combine(&2) => 3, i32
-        test_u8, 1.combine(&2) => 3, u8
-        test_u16, 1.combine(&2) => 3, u16
-        test_u32, 1.combine(&2) => 3, u32
-        test_usize, 1.combine(&2) => 3, usize
-        test_isize, 1.combine(&2) => 3, isize
-        test_f32, 1f32.combine(&2f32) => 3f32, f32
-        test_f64, 1f64.combine(&2f64) => 3f64, f64
-        test_option_i16, Some(1).combine(&Some(2)) => Some(3), Option<i16>
-        test_option_i16_none1, None.combine(&Some(2)) => Some(2), Option<i16>
-        test_option_i16_none2, Some(2).combine(&None) => Some(2), Option<i16>
+    semi_tests! {
+        test_i8, 1.combine(2) => 3, i8
+        test_i8_2, (&1).combine(&2) => 3, i8
+        test_i16, 1.combine(2) => 3, i16
+        test_i32, 1.combine(2) => 3, i32
+        test_u8, 1.combine(2) => 3, u8
+        test_u16, 1.combine(2) => 3, u16
+        test_u32, 1.combine(2) => 3, u32
+        test_usize, 1.combine(2) => 3, usize
+        test_isize, 1.combine(2) => 3, isize
+        test_f32, 1f32.combine(2f32) => 3f32, f32
+        test_f64, 1f64.combine(2f64) => 3f64, f64
+        test_option_i16, Some(1).combine(Some(2)) => Some(3), Option<i16>
+        test_option_i16_none1, None.combine(Some(2)) => Some(2), Option<i16>
+        test_option_i16_none2, Some(2).combine(None) => Some(2), Option<i16>
     }
 
     #[test]
-    fn test_string() {
-        let v1 = String::from("Hello");
-        let v2 = String::from(" world");
-        assert_eq!(v1.combine(&v2), "Hello world")
+    fn test_combine_hlist() {
+        let h1 = hlist![Some(1), 3.3, 53i64, "hello".to_owned()];
+        let h2 = hlist![Some(2), 1.2, 1i64, " world".to_owned()];
+        let h3 = hlist![Some(3), 4.5, 54, "hello world".to_owned()];
+        assert_eq!(h1.combine(h2), h3)
     }
 
     #[test]
-    fn test_vec_i32() {
+    fn test_combine_hlist_2() {
+        let h1 = hlist![Some(1), 3.3, 53i64, "hello"];
+        let h2 = hlist![Some(2), 1.2, 1i64, " world"];
+        let h3 = hlist![Some(3), 4.5, 54, "hello world".to_owned()]; // sadly types don't line up otherwise
+        assert_eq!(h1.combine(h2), h3)
+    }
+
+    #[test]
+    fn test_combine_str() {
+        assert_eq!("hello".combine(" world"), "hello world")
+    }
+
+    #[test]
+    fn test_combine_string() {
+        assert_eq!(
+            "hello".to_string().combine(" world".to_string()),
+            "hello world"
+        )
+    }
+
+    #[test]
+    fn test_combine_str_chained() {
+        assert_eq!("hello".combine(" world").combine(" yay"), "hello world yay")
+    }
+
+    #[test]
+    fn test_combine_i32_chained() {
+        let i1: i32 = 50;
+        let i2: i32 = 51;
+        let i3: i32 = 52;
+        let c: i32 = (&i1).combine(&i2).combine(&i3);
+        assert_eq!(c, 153)
+    }
+
+    #[test]
+    fn test_combine_option_str() {
+        assert_eq!(
+            Some("hello").combine(Some(" world")),
+            Some("hello world".to_string())
+        )
+    }
+
+
+    #[test]
+    fn test_max() {
+        assert_eq!(Max(1).combine(Max(2)), Max(2));
+
+        let v = vec![Max(1), Max(2), Max(3)];
+        assert_eq!(combine_all_option(v), Some(Max(3)));
+    }
+
+    #[test]
+    fn test_min() {
+        assert_eq!(Min(1).combine(Min(2)), Min(1));
+
+        let v = vec![Min(1), Min(2), Min(3)];
+        assert_eq!(combine_all_option(v), Some(Min(1)));
+    }
+
+    #[test]
+    fn test_all() {
+        assert_eq!(All(3).combine(All(5)), All(1));
+        assert_eq!(All(true).combine(All(false)), All(false));
+    }
+
+    #[test]
+    fn test_any() {
+        assert_eq!(Any(3).combine(Any(5)), Any(7));
+        assert_eq!(Any(true).combine(Any(false)), Any(true));
+    }
+
+    #[test]
+    fn test_combine_all_option() {
         let v1 = vec![1, 2, 3];
-        let v2 = vec![4, 5, 6];
-        assert_eq!(v1.combine(&v2), vec![1, 2, 3, 4, 5, 6])
+        assert_eq!(combine_all_option(v1), Some(6));
+        let v2 = vec![Some(1), Some(2), Some(3)];
+        assert_eq!(combine_all_option(v2), Some(Some(6)));
     }
 
     #[test]
-    fn test_refcell() {
-        let v1 = RefCell::new(1);
-        let v2 = RefCell::new(2);
-        assert_eq!(v1.combine(&v2), RefCell::new(3))
+    fn test_combine_n() {
+        assert_eq!(combine_n(1, 3), 3);
+        assert_eq!(combine_n(2, 1), 2);
+        assert_eq!(combine_n(Some(2), 4), Some(8));
     }
 
     #[test]
@@ -439,46 +566,9 @@ mod tests {
         expected.insert(2);
         expected.insert(3);
         expected.insert(4);
-        assert_eq!(v1.combine(&v2), expected)
+        assert_eq!(v1.combine(v2), expected)
     }
 
-    #[test]
-    fn test_tuple() {
-        let t1 = (1, 2.5f32, String::from("hi"), Some(3));
-        let t2 = (1, 2.5f32, String::from(" world"), None);
-
-        let expected = (2, 5.0f32, String::from("hi world"), Some(3));
-
-        assert_eq!(t1.combine(&t2), expected)
-    }
-
-    #[test]
-    fn test_max() {
-        assert_eq!(Max(1).combine(&Max(2)), Max(2));
-
-        let v = vec![Max(1), Max(2), Max(3)];
-        assert_eq!(combine_all_option(&v), Some(Max(3)));
-    }
-
-    #[test]
-    fn test_min() {
-        assert_eq!(Min(1).combine(&Min(2)), Min(1));
-
-        let v = vec![Min(1), Min(2), Min(3)];
-        assert_eq!(combine_all_option(&v), Some(Min(1)));
-    }
-
-    #[test]
-    fn test_all() {
-        assert_eq!(All(3).combine(&All(5)), All(1));
-        assert_eq!(All(true).combine(&All(false)), All(false));
-    }
-
-    #[test]
-    fn test_any() {
-        assert_eq!(Any(3).combine(&Any(5)), Any(7));
-        assert_eq!(Any(true).combine(&Any(false)), Any(true));
-    }
 
     #[test]
     fn test_hashmap() {
@@ -493,29 +583,14 @@ mod tests {
         expected.insert(1, Some("Hello World".to_owned()));
         expected.insert(2, Some("Goodbye".to_owned()));
         expected.insert(4, Some("Nope".to_owned()));
-        assert_eq!(v1.combine(&v2), expected)
+        assert_eq!(v1.combine(v2), expected)
     }
 
     #[test]
-    fn test_combine_all_option() {
-        let v1 = &vec![1, 2, 3];
-        assert_eq!(combine_all_option(v1), Some(6));
-        let v2 = &vec![Some(1), Some(2), Some(3)];
-        assert_eq!(combine_all_option(v2), Some(Some(6)));
+    fn test_refcell() {
+        let v1 = RefCell::new(1);
+        let v2 = RefCell::new(2);
+        assert_eq!(v1.combine(v2), RefCell::new(3))
     }
 
-    #[test]
-    fn test_combine_n() {
-        assert_eq!(combine_n(&1, 3), 3);
-        assert_eq!(combine_n(&2, 1), 2);
-        assert_eq!(combine_n(&Some(2), 4), Some(8));
-    }
-
-    #[test]
-    fn test_combine_hlist() {
-        let h1 = hlist![Some(1), 3.3, 53i64, "hello".to_owned()];
-        let h2 = hlist![Some(2), 1.2, 1i64, " world".to_owned()];
-        let h3 = hlist![Some(3), 4.5, 54, "hello world".to_owned()];
-        assert_eq!(h1.combine(&h2), h3)
-    }
 }
